+++ v1/opt/splunk/setup/leader/002-defaults.sh
@@ -4,11 +4,11 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 source $DIR/../../../../lib/helpers.sh
 source /etc/environment
 # logging-config defaults
-etcd-set /splunk/config/label-index "logging-index"
-etcd-set /splunk/config/label-type "logging-type"
-etcd-set /splunk/config/label-sourcetype "logging-sourcetype"
-etcd-set /splunk/config/label-volume "logging-volume"
-etcd-set /splunk/config/heavyforwarder-endpoint "https://localhost:8089"
+etcd-set /splunk/config/label-index "index"
+etcd-set /splunk/config/label-type "type"
+etcd-set /splunk/config/label-sourcetype "source_type"
+etcd-set /splunk/config/label-volume "volume"
+etcd-set /splunk/config/heavyforwarder-endpoint "http://localhost:8089"
 etcd-set /splunk/config/heavyforwarder-auth "admin:changeme"
 etcd-set /splunk/config/logging-config/poller-frequency 10
 # secops defaults
@@ -19,13 +19,15 @@ etcd-set /splunk/config/secops/forwarder-cert ""
 etcd-set /splunk/config/secops/ca-cert ""
 etcd-set /splunk/config/secops/index "os"
 etcd-set /splunk/config/secops/sourcetype "syslog"
-etcd-set /splunk/config/secops/rootca-format "crt"
+etcd-set /splunk/config/secops/rootca-format "pem"
 etcd-set /splunk/config/secops/certpath-format "crt"
 # cloudops defaults
 etcd-set /splunk/config/cloudops/enable-forwarder 0
 etcd-set /splunk/config/cloudops/forwarder-cert ""
 etcd-set /splunk/config/cloudops/ca-cert ""
 etcd-set /splunk/config/cloudops/forwarder-server-list ""
+etcd-set /splunk/config/cloudops/hvc-endpoint ""
+etcd-set /splunk/config/cloudops/lvc-endpoint ""
 etcd-set /splunk/config/cloudops/sslpassword ""
 etcd-set /splunk/config/cloudops/index "ethos-sandbox-ue1"
 etcd-set /splunk/config/cloudops/sourcetype "journald"
@@ -52,5 +54,3 @@ etcd-set /splunk/config/heavyforwarder/scrub-syslog 0
 
 # make logging elb availible to etcd
 etcd-set /environment/LOGGING_ELB $FLUENTD_INTERNAL_ELB
-
-